Rename learningstep forms to include step in name
